refactor: 절대 경로가 빠진 파일 수정
